# docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SignVault API
  version: 0.2.0
  description: |
    Kontrak API SignVault v0.2 (sumber kebenaran, prd.md §10).
    Scope: backend (apps/backend, packages/database, docs/api-docs).
    v0.2: Google OAuth (POST /auth/google), onboarding (PATCH
    /users/me/complete-profile), CAPTCHA Turnstile, rate-limit auth 10/mnt.
    Base path: /api/v1.
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev (Uvicorn native, tanpa Docker)
...
paths:
...
  /sign-requests/pending:
    get:
      tags: [sign-requests]
      summary: Daftar request menunggu tanda tangan (Signer)
      operationId: listPendingSignRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Daftar pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignRequestPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sign-requests/{id}/approve:
    post:
      tags: [sign-requests]
      summary: Signer menyetujui + menandatangani satu dokumen (QR tertempel)
      operationId: approveSignRequest
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '201':
          description: Signature terbuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signature'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /sign-requests/batch-approve:
    post:
      tags: [sign-requests]
      summary: Batch signing (transaksi + optimistic locking)
      operationId: batchApproveSignRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchApproveRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Hasil per item (sukses/gagal 개별, tidak setengah jalan)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchApproveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sign-requests/{id}/reject:
    post:
      tags: [sign-requests]
      summary: Signer menolak dokumen beserta alasan
      operationId: rejectSignRequest
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Ditolak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /verify/{sig_id}:
    get:
      tags: [verify]
      summary: Verifikasi publik via QR/link tanpa login
      operationId: verifyBySigId
      security: []
      parameters:
        - name: sig_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hasil verifikasi VALID / TIDAK VALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResult'
  /verify/upload:
    post:
      tags: [verify]
      summary: Verifikasi publik via upload PDF tanpa login
      operationId: verifyUpload
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Hasil verifikasi VALID / TIDAK VALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /audit-logs:
    get:
      tags: [audit-logs]
      summary: 'Audit log (Super Admin, Org Admin). Immutable: tanpa POST/PUT/DELETE'
      operationId: listAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Daftar audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Token hilang / tidak valid / kedaluwarsa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Role tidak diizinkan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ValidationError:
      description: Validasi request gagal (tipe/format field salah)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: Konflik status (mis. sudah di-sign / double request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Token tidak valid atau kedaluwarsa.
...
    KeyAlgorithm:
      type: string
      enum: [RSA_PSS_2048, ECDSA_P256, ED25519]
...
    SignRequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    SignRequest:
      type: object
      required: [id, documentId, signerId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        signerId:
          type: string
        status:
          $ref: '#/components/schemas/SignRequestStatus'
        message:
          type: string
          nullable: true
        rejectReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApproveRequest:
      type: object
      required: [keyPairId]
      properties:
        keyPairId:
          type: string
        position:
          type: string
          maxLength: 100
          description: Jabatan signer saat signing (opsional, snapshot ke QR)
    BatchApproveRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [signRequestId, keyPairId]
            properties:
              signRequestId:
                type: string
              keyPairId:
                type: string
              position:
                type: string
                maxLength: 100
                description: Jabatan signer saat signing (opsional)
    BatchApproveItemResult:
      type: object
      required: [signRequestId, success]
      properties:
        signRequestId:
          type: string
        success:
          type: boolean
        signatureId:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorBody'
          nullable: true
    BatchApproveResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/BatchApproveItemResult'
    RejectRequest:
      type: object
      required: [rejectReason]
      properties:
        rejectReason:
          type: string
    Signature:
      type: object
      required: [id, documentId, signerId, keyPairId, algorithm, signatureValue, signedHash, canonicalMetadata, qrPayload, createdAt]
      properties:
        id:
          type: string
          description: sig_id untuk verifikasi publik
        documentId:
          type: string
        signerId:
          type: string
        keyPairId:
          type: string
        algorithm:
          $ref: '#/components/schemas/KeyAlgorithm'
        signatureValue:
          type: string
          description: Base64 signature atas SHA-256(PDF) + metadata kanonis
        signedHash:
          type: string
        canonicalMetadata:
          type: string
        qrPayload:
          type: string
          description: >-
            JSON publik ringkas {v,doc,sig,name,pos,org,at,alg,url} (maks 300 char,
            tanpa data sensitif). Hanya rujukan — keaslian dari verifikasi kriptografis.
        signerPosition:
          type: string
          nullable: true
          description: Jabatan snapshot saat signing
        signedPdfPath:
          type: string
          nullable: true
        sigFormat:
          type: string
          enum: [LEGACY, PADES]
          description: PADES = ByteRange RSA-PSS tertanam (Adobe-verifiable)
        byteRange:
          type: string
          nullable: true
          description: ByteRange PAdES "[a b c d]"
        createdAt:
          type: string
          format: date-time
    VerifyResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [VALID, INVALID]
        documentName:
          type: string
        signerName:
          type: string
        signedAt:
          type: string
          format: date-time
        reason:
          type: string
        auditTrail:
          type: array
          items:
            type: object
            required: [event, at]
            properties:
              event:
                type: string
              at:
                type: string
                format: date-time
              actor:
                type: string
    AuditLog:
      type: object
      required: [id, action, details, createdAt]
      properties:
        id:
          type: string
        actorId:
          type: string
          nullable: true
        action:
          type: string
        entity:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
        details:
          type: object
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    SignRequestPage:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SignRequest'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    AuditLogPage:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...
security:
  - bearerAuth: []